Updated wording of the Highcharts attribution in the readme and license files
Update was based on Mathew Vis-Dunbar's suggestion:

The deployment of Curriculum MAP on UBC servers is covered by the CC licence granted by Highcharts. The sharing of the code for the project on GitHub is not covered directly by a licence from Highcharts, just an ok that since GitHub is not a deployment platform, but a development platform, you’re (the team) is allowed to post Highcharts code there. So I’d suggest editing the text to indicate that in order to use the Highcharts functionality if someone chooses to fork your project, they will need to acquire a licence from Highcharts.