Fixed BC Degree Standards Bug + Course # Bug
BC Degree Standards charts were broken at course level because the syntax used to retrieve the Standard Outcome Maps changed during the Laravel update. I also corrected a bug in the Program Overview which was caused by courses not having a course number, I just added an if statement and an extra case.